modified:   JAR-EXECUTABLE/jeux-zone-geographique.jar 	modified:   bin/.gitignore 	deleted:    bin/characteres/Charactere.class 	deleted:    bin/characteres/Intrus.class 	deleted:    bin/characteres/Robot.class 	modified:   bin/interfaceGraphique/FenetreChoixEmplacement.class 	modified:   bin/interfaceGraphique/FenetreJeux.class 	modified:   bin/interfaceGraphique/FenetreResultat.class 	modified:   bin/zoneGeographique/ZoneGeographique.class 	renamed:    src/characteres/Charactere.java -> src/caracteres/Caractere.java 	renamed:    src/characteres/Intrus.java -> src/caracteres/Intrus.java 	renamed:    src/characteres/Robot.java -> src/caracteres/Robot.java 	modified:   src/interfaceGraphique/FenetreChoixEmplacement.java 	modified:   src/interfaceGraphique/FenetreJeux.java 	modified:   src/interfaceGraphique/FenetreResultat.java 	modified:   src/zoneGeographique/ZoneGeographique.java

+++ src/interfaceGraphique/FenetreResultat.java
@@ -50,7 +50,6 @@ public class FenetreResultat extends JFrame implements ActionListener {
 		} else if (zoneGeo.get_num_gagnant() == 2) {
 			infoLabel.setText("Vous avez vole " + zoneGeo.getNbArgentVole() + " sacs d argent");
 		}
-
 		JLabel dureeJeuxLabel = new JLabel("Duree du jeux : " + format_duree(duree));
 		dureeJeuxLabel.setBounds(40, 100, 400, 30);
 		dureeJeuxLabel.setFont(new Font("Aerial", Font.BOLD, 15));
@@ -146,10 +145,10 @@ public class FenetreResultat extends JFrame implements ActionListener {
 	public String format_duree(double duree) {
 		String s;
 		if (duree < 60) {
-			s = String.valueOf(duree) + " secondes";
+			s = String.valueOf((int)duree) + " secondes";
 		} else {
 			int min = (int) (duree / 60);
-			double sec = duree - 60 * min;
+			int sec =  (int) duree - 60 * min;
 			s = String.valueOf(min) + " minutes et " + String.valueOf(sec) + " secondes";
 		}
 		return s;
